refactor: move command errors into their own files
