+++ src/react/loom-app/loom-state-provider/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { LoomState } from "src/shared/loom-state/types/loom-state";
-import LoomStateCommand from "src/shared/loom-state/loom-state-command";
+import LoomStateCommand from "src/shared/loom-state/commands/loom-state-command";
 import { useLogger } from "src/shared/logger";
 import RowSortCommand from "src/shared/loom-state/commands/row-sort-command";
 import { useAppMount } from "src/react/loom-app/app-mount-provider";
