+++ src/shared/loom-state/commands/cell-body-update-command.spec.ts
@@ -1,5 +1,5 @@
 import { createTestLoomState } from "src/shared/loom-state/loom-state-factory";
-import { CommandUndoError } from "./command-errors";
+import CommandUndoError from "./command-undo-error";
 import CellBodyUpdateCommand from "./cell-body-update-command";
 import { advanceBy, clear } from "jest-date-mock";
