+++ src/shared/loom-state/commands/cell-body-update-command.ts
@@ -1,6 +1,6 @@
 import { rowLastEditedTime, rowLastEditedTimeUpdate } from "../row-utils";
 import CellNotFoundError from "src/shared/error/cell-not-found-error";
-import LoomStateCommand from "../loom-state-command";
+import LoomStateCommand from "./loom-state-command";
 import { BodyCell, LoomState } from "../types/loom-state";
 
 export default class CellBodyUpdateCommand extends LoomStateCommand {
