+++ src/shared/loom-state/commands/cell-footer-update-command.ts
@@ -1,5 +1,5 @@
 import CellNotFoundError from "src/shared/error/cell-not-found-error";
-import LoomStateCommand from "../loom-state-command";
+import LoomStateCommand from "./loom-state-command";
 import { FooterCell, LoomState } from "../types/loom-state";
 
 export default class CellFooterUpdateCommand extends LoomStateCommand {
