+++ src/shared/loom-state/commands/cell-header-update-command.ts
@@ -1,5 +1,5 @@
 import CellNotFoundError from "src/shared/error/cell-not-found-error";
-import LoomStateCommand from "../loom-state-command";
+import LoomStateCommand from "./loom-state-command";
 import { HeaderCell, LoomState } from "../types/loom-state";
 
 export default class CellHeaderUpdateCommand extends LoomStateCommand {
