+++ src/shared/loom-state/commands/command-redo-error.ts
@@ -1,0 +1,6 @@
+export default class CommandRedoError extends Error {
+	constructor() {
+		super(`undo() must be called before redo() is available`);
+		this.name = "CommandRedoError";
+	}
+}
