+++ src/shared/loom-state/commands/command-undo-error.ts
@@ -1,0 +1,6 @@
+export default class CommandUndoError extends Error {
+	constructor() {
+		super(`execute() must be called before undo() is available`);
+		this.name = "CommandUndoError";
+	}
+}
