+++ src/shared/loom-state/commands/row-sort-command.ts
@@ -1,7 +1,7 @@
 import CellNotFoundError from "src/shared/error/cell-not-found-error";
 import RowNotFoundError from "src/shared/error/row-not-found-error";
 import TagNotFoundError from "src/shared/error/tag-not-found-error";
-import LoomStateCommand from "../loom-state-command";
+import LoomStateCommand from "./loom-state-command";
 import {
 	BodyCell,
 	BodyRow,
