+++ src/shared/loom-state/commands/tag-delete-command.ts
@@ -1,4 +1,4 @@
-import LoomStateCommand from "../loom-state-command";
+import LoomStateCommand from "./loom-state-command";
 import { LoomState, Tag } from "../types/loom-state";
 import TagNotFoundError from "src/shared/error/tag-not-found-error";
